Add files via upload
COVID-19 Global Analysis and Forecasting Project

This project focuses on analyzing and forecasting COVID-19 trends across the world using the most recent data. Key aspects of the analysis include:
 - Active cases analysis
 - Death cases analysis
 - Confirmed cases analysis
 - Recovered cases analysis

Additionally, it features:
 - Trend analysis for active, confirmed, death, and recovered cases
 - Time series forecasting for confirmed, active, death, and recovered cases

This project employs various data science techniques and models to help visualize and predict the spread of COVID-19 across the globe.